Separate base parts out of SystemLog
